+++ src/config.js
@@ -64,7 +64,7 @@ const ALLOWED_EVENT_OPTIONS_AD_HOC = [
   "checkForNextChunk",
   "retryFailedAfter",
   "propagateHeaders",
-  "propagatedContextProperties",
+  "propagateContextProperties",
   "retryOpenAfter",
   "multiInstanceProcessing",
   "kind",
@@ -518,7 +518,7 @@ class Config {
       ? Object.fromEntries(new Map(event.appInstances.map((a) => [a, true])))
       : null;
     event.propagateHeaders = event.propagateHeaders ?? [];
-    event.propagatedContextProperties = event.propagatedContextProperties ?? [];
+    event.propagateContextProperties = event.propagateContextProperties ?? [];
     event.retryFailedAfter = event.retryFailedAfter ?? DEFAULT_RETRY_AFTER;
     event.retryOpenAfter = event.retryOpenAfter ?? DEFAULT_RETRY_AFTER;
   }