+++ src/outbox/eventQueueAsOutbox.js
@@ -43,7 +43,7 @@ function outboxed(srv, customOpts) {
 
     const outboxOpts = config.getEventConfig(CDS_EVENT_TYPE, subType, srvConfig.namespace);
     const eventHeaders = getPropagatedHeaders(outboxOpts, req);
-    const contextProperties = getPropagatedContextProperties(outboxOpts, req);
+    const contextProperties = getPropagateContextProperties(outboxOpts, req);
     if (["persistent-outbox", "persistent-queue"].includes(outboxOpts.kind)) {
       await _mapToEventAndPublish(req, srvConfig.namespace, subType, eventHeaders, contextProperties);
       return;
@@ -81,8 +81,8 @@ const getPropagatedHeaders = (config, req) => {
   return Object.assign(propagateHeaders, req.headers);
 };
 
-const getPropagatedContextProperties = (config, req) => {
-  return config.propagatedContextProperties.reduce((properties, name) => {
+const getPropagateContextProperties = (config, req) => {
+  return config.propagateContextProperties.reduce((properties, name) => {
     if (name in req.tx.context) {
       properties[name] = req.tx.context[name];
     }